Move the utf8 column width to the printer
